// gui/src/app/shared/models/storage-device.ts
// Copyright 2019, Pure Storage Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//    http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

export class StorageDevice {
    id: string; // UUID
    name: string; // User-chosen display name
    mgmt_endpoint: string; // FlashArray and FlashBlade
    api_token: string;
    status: string; // This could be an enum: to be determined
    device_type: string;
    model: string;
    version: string;
    _as_of: string; // Should be Unix standard time format
    _last_updated: string; // Should be Unix standard time format
}